Fix crash when using auto address configuration. rdomain should be 32bit after the last change.

+++ iked/vroute-netlink.c
@@ -528,7 +528,7 @@ vroute_getcloneroute(struct iked *env, struct imsg *imsg)
 	struct sockaddr_storage	 mask;
 	uint8_t			*ptr;
 	size_t			 left;
-	uint8_t			 rdomain;
+	uint32_t		 rdomain;
 
 	ptr = (uint8_t *)imsg->data;
 	left = IMSG_DATA_SIZE(imsg);
